+++ src/app/chat/features/ChatInput/STT/index.tsx
@@ -1,10 +1,3 @@
-import {
-  OpenaiSpeechRecognitionOptions,
-  useOpenaiSTTWithPSR,
-  useOpenaiSTTWithSR,
-  usePersistedSpeechRecognition,
-  useSpeechRecognition,
-} from '@lobehub/tts';
 import { ActionIcon, Icon } from '@lobehub/ui';
 import { Button, Dropdown } from 'antd';
 import { createStyles } from 'antd-style';
@@ -14,9 +7,9 @@ import { memo, useCallback } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Flexbox } from 'react-layout-kit';
 
+import { useSTT } from '@/hooks/useSTT';
 import { settingsSelectors, useGlobalStore } from '@/store/global';
 import { useSessionStore } from '@/store/session';
-import { agentSelectors } from '@/store/session/selectors';
 
 const useStyles = createStyles(({ css, token }) => ({
   recording: css`
@@ -30,51 +23,21 @@ const useStyles = createStyles(({ css, token }) => ({
 const STT = memo<{ mobile?: boolean }>(({ mobile }) => {
   const { t } = useTranslation('chat');
   const { styles } = useStyles();
-  const ttsConfig = useSessionStore(agentSelectors.currentAgentTTS, isEqual);
+
   const settings = useGlobalStore(settingsSelectors.currentSettings, isEqual);
-  const [locale, openAIAPI, openAIProxyUrl] = useGlobalStore((s) => [
-    settingsSelectors.currentLanguage(s),
-    settingsSelectors.openAIAPI(s),
-    settingsSelectors.openAIProxyUrl(s),
-  ]);
+
   const [loading, updateInputMessage] = useSessionStore((s) => [
     !!s.chatLoadingId,
     s.updateInputMessage,
   ]);
 
   const isPersisted = settings.tts.sttPersisted;
-  const sttLocale =
-    ttsConfig?.sttLocale && ttsConfig.sttLocale !== 'auto' ? ttsConfig.sttLocale : locale;
 
-  let useSTT;
-  let options: any = {};
-
-  switch (settings.tts.sttServer) {
-    case 'openai': {
-      useSTT = isPersisted ? useOpenaiSTTWithPSR : useOpenaiSTTWithSR;
-      options = {
-        api: {
-          key: openAIAPI,
-          proxy: openAIProxyUrl,
-        },
-        model: settings.tts.openAI.sttModel,
-      } as OpenaiSpeechRecognitionOptions;
-      break;
+  const { start, isLoading, stop, formattedTime, time } = useSTT((text) => {
+    if (loading) stop();
+    if (text) {
+      updateInputMessage(text);
     }
-    case 'browser': {
-      useSTT = isPersisted ? usePersistedSpeechRecognition : useSpeechRecognition;
-      break;
-    }
-  }
-
-  const { start, isLoading, stop, formattedTime, time } = useSTT(sttLocale, {
-    ...options,
-    onTextChange: (text) => {
-      if (loading) stop();
-      if (text) {
-        updateInputMessage(text);
-      }
-    },
   });
 
   const icon = isLoading ? MicOff : Mic;
